Revert "Fix/native/cancel pin on passphrase (#12876)"
This reverts commit 265170e31fcb1d55efd8b3c98b4f04163195f7c5.
