+++ suite-native/app/package.json
@@ -61,7 +61,6 @@
         "@suite-native/module-settings": "workspace:*",
         "@suite-native/navigation": "workspace:*",
         "@suite-native/notifications": "workspace:*",
-        "@suite-native/passphrase": "workspace:*",
         "@suite-native/receive": "workspace:*",
         "@suite-native/settings": "workspace:*",
         "@suite-native/state": "workspace:*",
