+++ suite-native/app/tsconfig.json
@@ -55,7 +55,6 @@
         { "path": "../module-settings" },
         { "path": "../navigation" },
         { "path": "../notifications" },
-        { "path": "../passphrase" },
         { "path": "../receive" },
         { "path": "../settings" },
         { "path": "../state" },
