+++ suite-native/module-connect-device/src/components/ConnectDeviceScreenHeader.tsx
@@ -19,9 +19,7 @@ import {
     selectDeviceModel,
     removeButtonRequests,
     selectIsDeviceDiscoveryActive,
-    selectIsUnauthorizedPassphraseDevice,
 } from '@suite-common/wallet-core';
-import { cancelPassphraseAndSelectStandardDeviceThunk } from '@suite-native/passphrase';
 
 import { ConnectingTrezorHelp } from './ConnectingTrezorHelp';
 
@@ -43,7 +41,6 @@ export const ConnectDeviceScreenHeader = ({
     const { showAlert, hideAlert } = useAlert();
 
     const device = useSelector(selectDevice);
-    const isUnauthorizedPassphraseDevice = useSelector(selectIsUnauthorizedPassphraseDevice);
     const deviceModel = useSelector(selectDeviceModel);
     const isDiscoveryActive = useSelector(selectIsDeviceDiscoveryActive);
 
@@ -74,26 +71,11 @@ export const ConnectDeviceScreenHeader = ({
                             : 'ButtonRequest_PinEntry',
                 }),
             );
-
-            // If pin screen was triggered from opening passphrase wallet (which creates new instance) and is cancelled, we need to select standard wallet
-            if (isUnauthorizedPassphraseDevice) {
-                dispatch(cancelPassphraseAndSelectStandardDeviceThunk());
-            }
-
             if (navigation.canGoBack()) {
                 navigation.goBack();
             }
         }
-    }, [
-        device,
-        deviceModel,
-        dispatch,
-        hideAlert,
-        isDiscoveryActive,
-        isUnauthorizedPassphraseDevice,
-        navigation,
-        showAlert,
-    ]);
+    }, [device, deviceModel, dispatch, isDiscoveryActive, navigation, showAlert, hideAlert]);
 
     // Handle hardware back button press same as cancel button
     useEffect(() => {
