+++ suite-native/module-passphrase/src/components/PassphraseScreenHeader.tsx
@@ -1,7 +1,7 @@
 import { useDispatch } from 'react-redux';
 import { useState } from 'react';
 
-import { useNavigation } from '@react-navigation/native';
+import { useNavigation, useRoute } from '@react-navigation/native';
 
 import {
     BottomSheet,
@@ -36,6 +36,8 @@ const buttonWrapperStyle = prepareNativeStyle(() => ({
 }));
 
 export const PassphraseScreenHeader = () => {
+    const route = useRoute();
+
     const { applyStyle } = useNativeStyles();
 
     const navigation = useNavigation<NavigationProp>();
@@ -54,7 +56,15 @@ export const PassphraseScreenHeader = () => {
         });
     };
 
-    const toggleBottomSheet = () => setShouldShowWarningBottomSheet(!shouldShowWarningBottomSheet);
+    const handlePress = () => {
+        if (route.name === PassphraseStackRoutes.PassphraseConfirmOnTrezor) {
+            setShouldShowWarningBottomSheet(true);
+        } else {
+            handleClose();
+        }
+    };
+
+    const handleCloseBottomSheet = () => setShouldShowWarningBottomSheet(false);
 
     return (
         <ScreenHeaderWrapper>
@@ -64,9 +74,9 @@ export const PassphraseScreenHeader = () => {
                 colorScheme="tertiaryElevation1"
                 accessibilityRole="button"
                 accessibilityLabel="close"
-                onPress={toggleBottomSheet}
+                onPress={handlePress}
             />
-            <BottomSheet isVisible={shouldShowWarningBottomSheet} onClose={toggleBottomSheet}>
+            <BottomSheet isVisible={shouldShowWarningBottomSheet} onClose={handleCloseBottomSheet}>
                 <VStack justifyContent="center" alignItems="center" padding="small" spacing="large">
                     <Text variant="titleSmall" textAlign="center">
                         <Translation id="modulePassphrase.confirmOnDevice.warningSheet.title" />
@@ -75,7 +85,7 @@ export const PassphraseScreenHeader = () => {
                         <Button colorScheme="redBold" onPress={handleClose}>
                             <Translation id="modulePassphrase.confirmOnDevice.warningSheet.primaryButton" />
                         </Button>
-                        <Button colorScheme="redElevation0" onPress={toggleBottomSheet}>
+                        <Button colorScheme="redElevation0" onPress={handleCloseBottomSheet}>
                             <Translation id="modulePassphrase.confirmOnDevice.warningSheet.secondaryButton" />
                         </Button>
                     </VStack>
