+++ suite-native/passphrase/src/index.ts
@@ -1,4 +1,3 @@
 export * from './passphraseSlice';
 export * from './passphraseThunks';
 export * from './useHandleDuplicatePassphrase';
-export * from './useHandleDeviceRequestsPassphrase';
