+++ suite-native/passphrase/src/passphraseSlice.ts
@@ -7,11 +7,8 @@ import {
     createDeviceInstanceThunk,
 } from '@suite-common/wallet-core';
 
-import { verifyPassphraseOnEmptyWalletThunk } from './passphraseThunks';
-
 type PassphraseState = {
     error: AuthorizeDeviceError | CreateDeviceInstanceError | null;
-    isVerifyingPassphraseOnEmptyWallet: boolean;
 };
 
 type PassphraseRootState = {
@@ -20,7 +17,6 @@ type PassphraseRootState = {
 
 const passphraseInitialState: PassphraseState = {
     error: null,
-    isVerifyingPassphraseOnEmptyWallet: false,
 };
 
 export const passphraseSlice = createSlice({
@@ -29,18 +25,6 @@ export const passphraseSlice = createSlice({
     reducers: {},
     extraReducers: builder => {
         builder
-            .addCase(verifyPassphraseOnEmptyWalletThunk.pending.type, state => {
-                state.isVerifyingPassphraseOnEmptyWallet = true;
-            })
-            .addMatcher(
-                isAnyOf(
-                    verifyPassphraseOnEmptyWalletThunk.fulfilled,
-                    verifyPassphraseOnEmptyWalletThunk.rejected,
-                ),
-                state => {
-                    state.isVerifyingPassphraseOnEmptyWallet = false;
-                },
-            )
             .addMatcher(
                 isAnyOf(authorizeDeviceThunk.pending, createDeviceInstanceThunk.pending),
                 state => {
@@ -66,8 +50,5 @@ export const selectPassphraseError = (state: PassphraseRootState) => state.passp
 export const selectPassphraseDuplicateError = (state: PassphraseRootState) => {
     return state.passphrase.error?.error === 'passphrase-duplicate' ? state.passphrase.error : null;
 };
-
-export const selectIsVerifyingPassphraseOnEmptyWallet = (state: PassphraseRootState) =>
-    state.passphrase.isVerifyingPassphraseOnEmptyWallet;
 
 export const passphraseReducer = passphraseSlice.reducer;
